requirements.txt
GitPython==3.1.31
requests==2.28.1
sqlite-utils==3.33
tqdm==4.65.0